Leanpub Book Information

Leanpub Book Information

lp_book_info(slug, api_key = NULL, secure = TRUE, verbose = TRUE,
  ...)

lp_summary(slug, api_key = NULL, secure = TRUE, verbose = TRUE, ...)

Arguments

slug
slug of the project

api_key
API key for Leanpub, passed to lp_api_key

secure
passed to lp_base_url for https

verbose
print diagnostic messages

...
additional options to pass to GET

Value

List of the result of the GET call and the content

Note

See https://leanpub.com/help/api

Examples

if (lp_have_api_key()) {
stat = lp_book_info(slug = "modernscientist")
}

#> GET command is:

#> {"id":33432,"title":"How to be a modern scientist","subtitle":null,"slug":"modernscientist","about_the_book":"\u003cp\u003eThe face of academia is changing. It is no longer sufficient to just publish or perish. We are now in an era where Twitter, Github, Figshare, and Alt Metrics are regular parts of the scientific workflow. Here I give high level advice about which tools to use, how to use them, and what to look out for. This book is appropriate for scientists at all levels who want to stay on top of the current technological developments affecting modern scientific careers. The book is based in part on the author's popular guides including guides for\u003c/p\u003e\n\u003cul\u003e\n\u003cli\u003e\u003ca href=\"https://github.com/jtleek/reviews\" target=\"_self\"\u003eReviewing papers\u003c/a\u003e\u003c/li\u003e\n\u003cli\u003e\u003ca href=\"https://github.com/jtleek/readingpapers\" target=\"_self\"\u003eReading papers\u003c/a\u003e\u003c/li\u003e\n\u003cli\u003e\u003ca href=\"https://github.com/jtleek/careerplanning\" target=\"_self\"\u003eCareer planning\u003c/a\u003e\u003c/li\u003e\n\u003cli\u003e\u003ca href=\"https://github.com/jtleek/talkguide\" target=\"_self\"\u003eGiving talks\u003c/a\u003e\u003c/li\u003e\n\u003c/ul\u003e\n\u003cp\u003eThe book is probably most suited to graduate students and postdocs in the sciences, but may be of interest to others who want to adapt their scientific process to use modern tools. \u003c/p\u003e\n\u003cp\u003e\u003c/p\u003e\n\u003cp\u003e\u003cem\u003eAbout the author: \u003ca href=\"http://www.jtleek.com\" target=\"_self\"\u003eJeff Leek\u003c/a\u003e is an Associate Professor of Biostatistics at the Johns Hopkins Bloomberg School of Public Health. He is a co-founder and co-director of the \u003ca href=\"https://www.coursera.org/specializations/jhu-data-science\" target=\"_self\"\u003eJohns Hopkins Data Science Specialization\u003c/a\u003e on Coursera that has enrolled over 3 million aspiring data scientists. His research has helped contributed to our understanding of the genomic basis of brain development, blunt force trauma, and cancer. He is blogs at \u003ca href=\"http://simplystatistics.org/\" target=\"_self\"\u003eSimply Statistics\u003c/a\u003e and can be found on twitter at \u003ca href=\"https://twitter.com/jtleek\" target=\"_self\"\u003e@jtleek\u003c/a\u003e and \u003ca href=\"https://twitter.com/simplystats\" target=\"_self\"\u003e@simplystats\u003c/a\u003e\u003c/em\u003e\u003c/p\u003e","last_published_at":"2016-04-08T17:59:03.000Z","page_count_published":82,"total_copies_sold":23954,"meta_description":"A book on how to be a scientist the modern way. ","author_string":"Jeffrey Leek","url":"http://leanpub.com/modernscientist","title_page_url":"https://s3.amazonaws.com/titlepages.leanpub.com/modernscientist/original?1518072265","minimum_price":"0.0","suggested_price":"10.0","image":"https://s3.amazonaws.com/titlepages.leanpub.com/modernscientist/medium?1518072265"}